moving types to /fleet/microsoft_mdm.go file
